fix: address Batch 6 & 7 CodeRabbit review (security, cleanup, tests)
- DonationForm: fix unused userLoading variable

- Tests: consolidate setup, fix webhook test structure & concurrent mocks

- Security: strict hex validation for signatures

- Cleanup: prettier formatting, remove dead code/imports

- Validation: ensure amounts are integers in input schema

+++ plugins/Razorpay/api/graphql/inputs.ts
@@ -46,7 +46,7 @@ export const RazorpayOrderInput = builder
     fields: (t) => ({
       organizationId: t.string({ required: false }),
       userId: t.string({ required: false }),
-      amount: t.float({ required: true }),
+      amount: t.int({ required: true }),
       currency: t.string({ required: true }),
       donorName: t.string({ required: false }),
       donorEmail: t.string({ required: false }),
